requirements.txt
python-binance

# python main.py --api-key "" --api-secret "" --symbol BTCUSDT --side BUY --qty 0.001

# export BINANCE_API_KEY=""
# export BINANCE_API_SECRET=""